ObJulia: Return or print output, with mimetype
* julia/init.jl: Certain backends (ess-julia) track the Julia process using its
stdout, and others (julia-snail) through callbacks provided by a mediating
layer.  To account for this, ObJulia can either return or print the
async-tracking uuid depending on an optional parameter `print_output'.
Additionally, it returns the mimetype used when writing the output file.  This
is needed for better result handling on the Org side.