ObJulia: Guess mimetype heuristically
* julia/packages.jl (package_mimes): lists mime types used by
supported packages (Plots and Gadfly) for now.

* julia/init.jl (auto_determine_mime, org_eval, output_mime): Guess
using `showable` if an evaluation result is of a mimetype supported by
packages like Plots and Gadfly. If true, use this mimetype instead of
the specified output file extension when writing to it. Note: this
behavior can be undesirable and will need to be controlled by a user
option.