ob-julia: Handle the link result format correctly
* ob-julia.el (org-babel-execute): When specifying the result
format as `link' along with the `file' result type, no output
should be written to the specified file since the code block is
supposed to write to the file instead
(see `(info "(org) Results of Evaluation")').  Fix this by
not using the specified file for babel output if the `link' or
`graphics' result options are specified.