ob-julia: Implement latexify header-arg
* ob-julia.el (org-babel-execute, org-babel-julia--maybe-make-raw,
ob-julia-latexify-star-environments,
org-babel-julia-process-results,
ob-julia-insert-latex-environment-advice): The `latexify' header
arg is now fully supported.  When used, the result type is forced
to `raw', and the result is processed further according to the the
options `org-babel-julia-force-latex-environment' and
`ob-julia-latexify-star-environments', which see.

TODO: Other forms of pre- and post-processing might be needed in
the future by other packages.  Should we add a pre/post processing
hook to ob-julia?